# openapi.yaml
openapi: 3.0.0
info:
  title: 'Star Citizen API'
  contact:
    email: foxftw@star-citizen.wiki
  version: 3.0.0
servers:
  -
    url: 'https://api.star-citizen.wiki'
paths:
...
  /api/starsystems:
    get:
      tags:
        - Starmap
        - RSI-Website
      summary: 'Starmap Starsystems Overview'
      description: 'Returns paginated starsystems, optionally including related resources.'
      operationId: dfb10343a9a3f85578e98f63dbfa7bf2
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
        -
          name: 'filter[affiliation]'
          in: query
          schema:
            type: string
        -
          name: 'filter[status]'
          in: query
          schema:
            type: string
        -
          name: 'filter[type]'
          in: query
          schema:
            type: string
        -
          name: 'filter[size]'
          in: query
          schema:
            type: number
        -
          name: include
          in: query
          description: 'Include additional relationships (affiliation, celestialObjects).'
          explode: false
          allowReserved: true
          schema:
            type: string
      responses:
        '200':
          description: 'List of Starsystems'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/starsystem'
  '/api/starsystems/{code}':
    get:
      tags:
        - Starmap
        - RSI-Website
      summary: 'Starsystem Detail'
      description: 'Retrieve a starsystem by code or identifier, with optional includes.'
      operationId: 1b478c502ef1afe25fbdcc192a8b7fd6
      parameters:
        -
          name: code
          in: path
          required: true
          schema:
            description: 'Starsystem code or identifier (e.g. SOL)'
            type: string
        -
          name: include
          in: query
          description: 'Include additional relationships (affiliation, celestialObjects).'
          explode: false
          allowReserved: true
          schema:
            type: string
      responses:
        '200':
          description: 'A Starsystem'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/starsystem'
        '404':
          description: 'No Starsystem with specified code found.'
  /api/starsystems/search:
    post:
      tags:
        - Starmap
        - RSI-Website
        - Search
      summary: 'Starsystem Search'
      description: 'Search for starsystems by code, cig_id, or name.'
      operationId: ae8a21d883ee8e48ac7be01e3ed3084a
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
      requestBody:
        description: 'Partial starsystem code or name to search for'
        required: true
        content:
          application/json:
            schema:
              properties:
                query:
                  type: string
              type: object
            example: '{"query": "Sol"}'
      responses:
        '200':
          description: 'List of matching Starsystems'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/starsystem'
...
components:
  schemas:
...
    celestial_object:
      title: 'Celestial Object'
      properties:
        id:
          type: integer
        code:
          type: string
        system_id:
          type: integer
        celestial_object_api_url:
          type: string
        name:
          type: string
        type:
          type: string
        age:
          type: integer
        habitable:
          type: boolean
        fairchanceact:
          type: boolean
        appearance:
          type: string
        designation:
          type: string
        distance:
          type: number
          format: float
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        axial_tilt:
          type: number
          format: float
        orbit_period:
          type: number
          format: float
        info_url:
          type: string
        description:
          type: object
        sensor:
          properties:
            population:
              type: number
              format: float
            economy:
              type: number
              format: float
            danger:
              type: number
              format: float
          type: object
        size:
          type: number
          format: float
        parent_id:
          type: integer
        time_modified:
          type: string
      type: object
    jumppoint:
      title: Jumppoint
      description: 'A jumppoint from the starmap'
      properties:
        id:
          type: integer
        size:
          type: string
        direction:
          type: string
        entry:
          properties:
            id:
              type: integer
            system_id:
              type: string
            system_api_url:
              type: string
            celestial_object_api_url:
              type: string
            status:
              type: string
            code:
              type: string
            designation:
              type: string
          type: object
        exit:
          properties:
            id:
              type: integer
            system_id:
              type: string
            system_api_url:
              type: string
            celestial_object_api_url:
              type: string
            status:
              type: string
            code:
              type: string
            designation:
              type: string
          type: object
      type: object
    starsystem:
      title: Starsystem
      properties:
        id:
          type: integer
        code:
          type: string
        system_api_url:
          type: string
        name:
          type: string
        status:
          type: string
        type:
          type: string
        position:
          properties:
            x:
              type: number
              format: float
            'y':
              type: number
              format: float
            z:
              type: number
              format: float
          type: object
        frost_line:
          type: number
          format: float
        habitable_zone_inner:
          type: number
          format: float
        habitable_zone_outer:
          type: number
          format: float
        info_url:
          type: string
        description:
          type: object
        aggregated:
          properties:
            size:
              type: number
              format: float
            population:
              type: number
              format: float
            economy:
              type: number
              format: float
            danger:
              type: number
              format: float
            stars:
              type: integer
            planets:
              type: integer
            moons:
              type: integer
            stations:
              type: integer
          type: object
        time_modified:
          type: string
        celestial_object:
          properties:
            celestial_object:
              properties:
                data:
                  $ref: '#/components/schemas/celestial_object'
              type: object
              nullable: true
          type: object
          nullable: true
        jumppoints:
          properties:
            jumppoints:
              properties:
                data:
                  $ref: '#/components/schemas/jumppoint'
              type: object
              nullable: true
          type: object
          nullable: true
      type: object
...
  parameters:
...
    page:
      name: page
      in: query
      description: 'Page number for pagination (starts at 1).'
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    page_number:
      name: 'page[number]'
      in: query
      description: 'Page number for pagination (starts at 1).'
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    page_size:
      name: 'page[size]'
      in: query
      description: 'Number of results to return per page.'
      required: false
      schema:
        type: integer
        default: 30
        maximum: 200
        minimum: 1
...